Reorganize text module slightly
